Make sure we really are in half-duplex mode; also fix link status read
The full duplex/half duplex mode of the ENC28J60 is supposed to be set
automatically from the polarity of one of the LEDs (!) but according to
the chip errata this is unreliable.  Make sure the chip is properly in
half-duplex mode: clear MACON3_FULDPX and PHCON1_PDPXMD.

(The chip will never negotiate full-duplex mode; if you want to use it
you have to configure your network switch appropriately.  Most people
with consumer grade kit won't be able to do this!)

Also a fix for link-up detection: we wanted bitwise AND to test bit 0x04
of the byte, not logical AND!