Remove central application call, replace with function pointer
uIP had an assumption that there would only be one TCP-based and one
UDP-based applicaiton compiled in - these would #define UIP_APPCALL in
their header files.  Replace this system with a function pointer
per-connection and define UIP_APPCALL and UIP_UDP_APPCALL to indirect
through those.

This commit compiles, but doesn't include the code to initialise the
application function pointer when a new connection is started (either
because we received a connection setup or because we made an outgoing
connection).

The function pointer init code will involve changes to uip.c - it's
separated into a separate commit because I'm trying to avoid changing
uip.c as much as I possibly can.