Working support for multiple apps
uip.c now indirects through a function pointer in the connection state
structure for TCP connections when calling applications.  Support
initialising this either on uip_connect() (for outgoing connections) or
when an incoming connection arrives for an app that has called
uip_listen()

There may be times that uip.c wants to call an app with the connection
pointer set to NULL.  I'm not sure how to handle this yet!

Still todo: extend the interface in NanodeUIP.h to support initialising
multiple apps.