Fix buffer sizes in hello-world application
The hello-world application was buggy - it didn't null-terminate the
string it read from the user before calling strncpy() on it.  Fix that,
and extend the application to demonstrate that we really do have a tcp
connection that keeps track of where we are.

If you're trying this example out, try having multiple connections open
to the nanode at once - telnet [ipaddr] 1000 from several windows at once.