Retry read of MAC device until it works
Check the SAK bit returned by the MAC address chip after every byte is
transferred.  If at any point we don't receive a SAK, abort and try
again.  Note that this will now freeze on Nanodes without a working MAC
address chip (but this means we won't be using nonsense MAC addresses on
the network).

The timing in the MAC chip reading code is still a bit iffy!  Things are
improved by disabling interrupts but it's still not great.