Remove MAC chip reading code; add example sketch
The MAC chip reading code embedded in NanodeUIP was a bit of a hack.
There's a better library available now to access the MAC address chip -
see https://github.com/sde1000/NanodeUNIO

Unfortunately, referring to a locally installed Arduino library from
another library seems not to be possible!

So the init() method now takes a pointer to the desired MAC address as a
parameter.  I've included an example sketch (examples -> uiptest) that
makes use of the NanodeUNIO library to set the correct MAC address while
initialising UIP.