Primitive conversion to C++
I've just dealt with compilation failures caused by incorrect types that
were ignored when these programs were compiled as C.  This version
compiles and runs, but results in a sketch that is a couple of hundred
bytes larger than before.

It should now be a lot easier to implement applications in the sketch
rather than having to add them to the library.