Separate hello-world app out into example sketch
The idea of having all apps be in the library was crazy, but was
necessary when the whole thing was a C++ wrapper around the C uip code.
Now it's much more sensible to have apps outside the library.

Some apps, like the DHCP client and resolver, will remain inside the
library.